Feat: 기획 클라이언트 데이터 반영 추가

+++ migration/data_manager.py
@@ -1,11 +1,12 @@
 import os
-import re
 import json
 import logging
 import shutil
+from dateutil import parser
 import pandas as pd
 from pathlib import Path
 from pandas import DataFrame
+import warnings
 
 
 class DataManager:
@@ -16,6 +17,7 @@ class DataManager:
             config = json.load(f)
         self._set_config(config)
         self._set_folder()
+        warnings.filterwarnings('ignore', category=UserWarning, module='openpyxl')
 
     def _set_config(self, config):
         self.PATH_FOR_ROOT = config['DEFAULT']['ROOT_DATA_DIR']
@@ -27,7 +29,8 @@ class DataManager:
 
     def _set_folder(self):
         # JSON 폴더 초기화
-        shutil.rmtree(self.PATH_FOR_JSON)
+        if Path(self.PATH_FOR_JSON).is_dir():
+            shutil.rmtree(self.PATH_FOR_JSON)
         # JSON 폴더 생성
         os.makedirs(self.PATH_FOR_DATA)
         os.makedirs(self.PATH_FOR_INFO)
@@ -37,29 +40,20 @@ class DataManager:
         # 행의 개수가 2보다 적으면 무시
         if df.shape[0] < 2:
             return df
+
         # 데이터 프레임 1행 추출
         mask_df = df.iloc[0]
+
         # 필터링할 컬럼값을 추출
         filtered = mask_df[mask_df.isin(targets)].keys()
         # 적용타입행과 필드타입행을 제외한 2행부터 JSON추출
-        return df[filtered].iloc[2:]
+        data_df = df[filtered].iloc[2:]
+        # 정의된 DB형식으로 데이터 포멧
+        self._translate_asdb(data_df, df)
+        return data_df
 
-    # EXCEL의 데이터가 다음과 같을때 필드 명 중복이 있을 경우
-    # DATAFRAME에 필드명이 reg_dt.1 로 들어오므로 .1을 자른다.
-    # ------------------------------------------------
-    # id     | name | reg_dt | reg_dt
-    # SERVER | SERVER | DATE |  SERVER
-    # 1  | test |  2022.04.09 | 2021-03-09T00:00:00
-    # ------------------------------------------------
-    def _translate_json(self, data_list: list):
-        match_str = r'\.\w'
-        for data in data_list:
-            str_match = [_ for _ in data.keys() if re.search(match_str, _)]
-            for key in str_match:
-                new_key = re.sub(match_str, '', key)
-                data[new_key] = data.pop(key)
-
-    def _save_json(self, df: DataFrame, save_path: str, file_name: str):
+    @staticmethod
+    def _save_json(df: DataFrame, save_path: str, file_name: str):
         # 행의 개수가 0이면 무시
         if df.shape[1] == 0:
             return
@@ -67,13 +61,57 @@ class DataManager:
         # Data frame을 JSON으로 변환
         json_data = df.to_json(orient='records')
         json_list = json.loads(json_data)
-        self._translate_json(json_list)
 
         # 지정한 경로로 Json파일 저장
         save_path = save_path + "/" + file_name + ".json"
         with open(save_path, "w", encoding='utf-8') as f:
             json.dump(json_list, f, ensure_ascii=False, indent=4, sort_keys=True, separators=(',', ': '))
         logging.info("Json 파일 저장 성공 : " + save_path)
+
+    # --------------------------------------------------------------
+    # EXCEL의 데이터가 다음과 같을때 1번째 행은 디비속성 이다.
+    # 디비속성 값으로 데이터를 포멧팅한다.
+    # --------------------------------------------------------------
+    # id     | name | reg_dt | reg_dt
+    # 0 : SERVER | SERVER | DATE |  SERVER
+    # 1 : int | string | datetime |  datetime <-- 디비속성
+    # 2 :  | test |  2022.04.09 | 2021-03-09T00:00:00
+    # --------------------------------------------------------------
+    # Ref :
+    # https://pandas.pydata.org/docs/reference/api/
+    def _translate_asdb(self, data_df: DataFrame, filter_df: DataFrame):
+        for col in data_df.columns:
+            for i in data_df.index:
+                field_type = filter_df.loc[1][col]
+                field_value = filter_df.loc[i][col]
+                data_df.loc[i][col] = self._value_astype(field_type, field_value)
+
+    def _value_astype(self, column_type: str, column_value: any):
+        try:
+            if column_type == "string":
+                return str(column_value)
+            elif column_type == "float":
+                return float(column_value)
+            elif column_type == "int":
+                return int(column_value)
+            elif column_type == "long":
+                return int(column_value)
+            elif column_type == "datetime":
+                return self.iso8601(column_value)
+            else:
+                return str(column_value)
+        except Exception as e:
+            logging.warning(str(f"Column[{column_value}]" + str(e)))
+            return str(e)
+
+    @staticmethod
+    def iso8601(date_text: str) -> str:
+        try:
+            date = parser.parse(str(date_text))
+            return date.astimezone().isoformat()
+        except Exception as e:
+            logging.warning(str(e))
+            return str(e)
 
     def excel_to_json(self):
         # Excel파일 가져오기
@@ -82,6 +120,7 @@ class DataManager:
             try:
                 # 첫번째 시트를 JSON 타겟으로 설정
                 df = pd.read_excel(excel, sheet_name=0)
+
                 # 파일 이름으로 JSON 파일 저장 : DATA
                 self._save_json(self._get_filtered(df, ['ALL', 'SERVER']), self.PATH_FOR_DATA, excel.stem)
                 # 파일 이름으로 JSON 파일 저장 : INFO
